backend/requirements.txt
python-dotenv
pydantic
langchain
langchain-google-genai
langchain-core
langchain-community
langchain-text-splitters
langgraph
pymupdf
faiss-cpu
fastapi
uvicorn